Split logic around around bus optimization between post and dispatch
The goal is to avoid injecting a `PersistentBus` in place where it's not needed.
