+++ util/src/main/java/org/killbill/billing/util/optimizer/BusDispatcherOptimizerNoop.java
@@ -17,8 +17,12 @@
 
 package org.killbill.billing.util.optimizer;
 
-import org.killbill.bus.api.PersistentBus;
+import org.killbill.bus.api.BusEvent;
 
-public interface BusOptimizer extends PersistentBus {
+public class BusDispatcherOptimizerNoop implements BusDispatcherOptimizer {
 
+    @Override
+    public boolean shouldDispatch(final BusEvent event) {
+        return true;
+    }
 }
