+++ util/src/test/java/org/killbill/billing/GuicyKillbillTestNoDBModule.java
@@ -22,6 +22,9 @@ import org.killbill.billing.platform.api.KillbillConfigSource;
 import org.killbill.billing.platform.test.glue.TestPlatformModuleNoDB;
 import org.killbill.billing.util.glue.IDBISetup;
 import org.killbill.billing.util.glue.MemoryGlobalLockerModule;
+import org.killbill.billing.util.optimizer.BusDispatcherOptimizer;
+import org.killbill.billing.util.optimizer.BusDispatcherOptimizerNoop;
+import org.killbill.billing.util.optimizer.BusDispatcherOptimizerOn;
 import org.killbill.billing.util.optimizer.BusOptimizer;
 import org.killbill.billing.util.optimizer.BusOptimizerNoop;
 import org.killbill.billing.util.optimizer.BusOptimizerOn;
@@ -49,8 +52,10 @@ public class GuicyKillbillTestNoDBModule extends GuicyKillbillTestModule {
             super.configureBus();
             if (killbillFeatures.isBusOptimizationOn()) {
                 this.bind(BusOptimizer.class).to(BusOptimizerOn.class).asEagerSingleton();
+                this.bind(BusDispatcherOptimizer.class).to(BusDispatcherOptimizerOn.class).asEagerSingleton();
             } else {
                 this.bind(BusOptimizer.class).to(BusOptimizerNoop.class).asEagerSingleton();
+                this.bind(BusDispatcherOptimizer.class).to(BusDispatcherOptimizerNoop.class).asEagerSingleton();
             }
         }
     }
